requirements.txt
asgiref==3.10.0
attrs==25.4.0
dj-database-url==3.0.1
Django==5.2.7
django-environ==0.12.0
django-filter==24.3
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
Faker==37.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
whitenoise==6.11.0
